p2p/protocols: remove unnecessary tests
